requirements.txt
requests
asyncio
loguru
websockets_proxy
fake_useragent
websockets==12.0
pymongo
PySocks